refactor(*): fix pylint error `import-outside-toplevel` where import doesn't lead to circular import
