+++ src/_balder/controllers/vdevice_controller.py
@@ -3,11 +3,11 @@ from typing import Type, Dict, Union, TYPE_CHECKING
 
 import logging
 import inspect
+from _balder.vdevice import VDevice
 from _balder.controllers.base_device_controller import BaseDeviceController
 from _balder.exceptions import DeviceScopeError
 
 if TYPE_CHECKING:
-    from _balder.vdevice import VDevice
     from _balder.feature import Feature
 
 logger = logging.getLogger(__file__)
@@ -26,7 +26,6 @@ class VDeviceController(BaseDeviceController):
     def __init__(self, related_cls, _priv_instantiate_key):
         super().__init__()
 
-        from _balder.vdevice import VDevice
         # this helps to make this constructor only possible inside the controller object
         if _priv_instantiate_key != VDeviceController.__priv_instantiate_key:
             raise RuntimeError('it is not allowed to instantiate a controller manually -> use the static method '
